device_simulator/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
pydantic
PyYAML